Added K3s poller utilization
